# Don't inherit from Hash