Inheriting from Ruby base classes [is problematic](https://tenderlovemaking.com/2014/06/02/yagni-methods-are-killing-me.html) and should be avoided, in my opinion.

With the current implementation, we are actually setting instance variables on an instance of `Hash`…

~~I have a performance optimization for the Config class that depends on this commit, but due to the “The PR relates to *only* one subject” constraint, that’ll have to wait~~

The 2nd commit is a performance optimization of `RuboCop::Config#to_s` that was previously introduced in 5806b71f43a7c959afd42c970f82f25870105edf, but accidentally removed in 0a74efa4accc6557c8cafb63bb249b67c73c30bc. Including the 2nd commit into this PR violates the “The PR relates to *only* one subject” decree, but on the other hand it’s what legitimizes removing the Hash superclass.

The runtime improvements when running on ~2500 files:
```
Before, cold cache
bin/rubocop  119.72s user 4.35s system 99% cpu 2:05.31 total

After, cold cache
bin/rubocop  117.67s user 4.53s system 96% cpu 2:06.31 total

Before, warm cache
bin/rubocop  12.72s user 1.55s system 97% cpu 14.582 total

After, warm cache
bin/rubocop  3.51s user 1.08s system 97% cpu 4.691 total
```

As you see, warm cache runtime improves by 2/3. On smaller projects, of course, the benefits are smaller.

-----------------

Before submitting the PR make sure the following are checked:

* [x] Wrote [good commit messages][1].
* [x] Commit message starts with `[Fix #issue-number]` (if the related issue exists).
* [x] Used the same coding conventions as the rest of the project.
* [x] Feature branch is up-to-date with `master` (if not - rebase it).
* [x] Squashed related commits together.
* [ ] Added tests.
* [x] Added an entry to the [Changelog](../blob/master/CHANGELOG.md) if the new code introduces user-observable changes. See [changelog entry format](../blob/master/CONTRIBUTING.md#changelog-entry-format).
* [x] All tests are passing.
* [x] The new code doesn't generate RuboCop offenses.
* [ ] The PR relates to *only* one subject with a clear title
  and description in grammatically correct, complete sentences.

[1]: http://tbaggery.com/2008/04/19/a-note-about-git-commit-messages.html