@bquorning 2016-09-12
> I’d like some help with that failing test, please.

By making an assertion on `$stderr.string` I managed to figure out that `Config` needed a `#map` method.