@bquorning 2016-09-12
I didn’t add any tests. Is that necessary? `expect(RuboCop::Config.new).to_not be_a(Hash)` doesn’t add much value, I think.