@bquorning 2016-09-12
It’s trivial to remove `.[]`, `#==` and `#eql`, but should I include such a commit in this PR, or in a follow-up?