@bquorning 2016-09-15
Follow-up in #3497.